+++ src/lib/components/EntriesTable/EntriesTable.jsx
@@ -38,6 +38,7 @@ class EntriesTable extends Component {
               onChange={() => toggleSelect(entry)}
             />
           </td>
+          <td>{entry.id}</td>
           <td>
             <Link to={`/project/${projectId}/entries/${entry.entryTypeId}/${entry.id}`}>
               {entry.name}
@@ -45,6 +46,9 @@ class EntriesTable extends Component {
           </td>
           <td>
             {entry.tags.join(', ')}
+          </td>
+          <td>
+            {entry.entryType.id}
           </td>
           <td>
             {entry.entryType.name}
@@ -67,8 +71,10 @@ class EntriesTable extends Component {
         <thead>
           <tr>
             <th></th>
+            <th>ID</th>
             <th>Name</th>
             <th>Tags</th>
+            <th>Entry type ID</th>
             <th>Entry type</th>
             <th>Last modified by</th>
             <th>Published</th>