+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.8.2
+* Now showing ID and Entry type ID columns on entries table
+* Now showing ID on entry type list
+* App will now redirect to login screen when logging out on another open tab
+
 # 0.8.1
 * Fixing dependencies required by server.js